+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changed in 1.1.0
 ----------------
 
 - `Added` ScriptExecutor main class.
+- `Added` Added GZISOPEN / GZOSOPEN to ZipFunctions.
 - `Changed` Exception thrown in ScriptValue.createForType(...) if conversion absolutely fails.